requirements.txt
flask
boto3
gunicorn
python-dotenv
Werkzeug